requirements.txt
six
bitcoin-python
cryptokit
gevent
pyyaml
